requirements.txt
motor>=3.3.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain-core>=0.3.0

# Required for steering layer (RL replacement)
torch>=2.0.0
